1.54 Перевожу на компоуз экраны авторизации, загрузки изображений
