+++ app/src/main/java/com/pavlov/nearWarSecrets/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.pavlov.nearWarSecrets.ui.main
+package com.pavlov.nearWarSecrets
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -6,13 +6,15 @@ import androidx.activity.compose.setContent
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
 import com.pavlov.nearWarSecrets.navigation.NavGraph
-import com.pavlov.nearWarSecrets.ui.theme.MyShadowGalleryTheme
+import com.pavlov.nearWarSecrets.ui.theme.NearWarSecretsTheme
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            MyShadowGalleryTheme {
+            NearWarSecretsTheme {
                 val navController = rememberNavController()
                 NavGraph(
                     navController = navController,
