+++ app/src/main/java/com/pavlov/nearWarSecrets/data/model/BottomNavItem.kt
@@ -1,0 +1,5 @@
+package com.pavlov.nearWarSecrets.data.model
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class BottomNavItem(val title: String, val route: String, val icon: ImageVector)
