+++ app/src/main/java/com/pavlov/nearWarSecrets/ui/main/MainScreen.kt
@@ -9,14 +9,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.material.*
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Storage
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalContext
-import androidx.navigation.NavHostController
-import com.pavlov.nearWarSecrets.navigation.NavGraph
+import com.pavlov.nearWarSecrets.navigation.BottomNavigationBar
+import com.pavlov.nearWarSecrets.ui.itemLoader.ItemLoaderScreen
 
 @Composable
 fun MainScreen() {
@@ -24,43 +18,20 @@ fun MainScreen() {
     Scaffold(
         bottomBar = { BottomNavigationBar(navController) }
     ) { innerPadding ->
-        NavGraph(
+        NavHost(
             navController = navController,
-            activity = LocalContext.current as Activity,
+            startDestination = "item_loader",
             modifier = Modifier.padding(innerPadding)
-        )
-    }
-}
-
-@Composable
-fun BottomNavigationBar(navController: NavHostController) {
-    val items = listOf(
-        BottomNavItem("Item Loader", "item_loader", Icons.Default.Add),
-        BottomNavItem("Storage", "storage", Icons.Default.Storage),
-        BottomNavItem("Settings", "settings", Icons.Default.Settings)
-    )
-    BottomNavigation {
-        val navBackStackEntry by navController.currentBackStackEntryAsState()
-        val currentRoute = navBackStackEntry?.destination?.route
-        items.forEach { item ->
-            BottomNavigationItem(
-                icon = { Icon(item.icon, contentDescription = item.title) },
-                label = { Text(item.title) },
-                selected = currentRoute == item.route,
-                onClick = {
-                    navController.navigate(item.route) {
-                        navController.graph.startDestinationRoute?.let { route ->
-                            popUpTo(route) {
-                                saveState = true
-                            }
-                        }
-                        launchSingleTop = true
-                        restoreState = true
-                    }
-                }
-            )
+        ) {
+            composable("item_loader") {
+                ItemLoaderScreen()
+            }
+            composable("storage") {
+//                StorageScreen()
+            }
+            composable("settings") {
+//                SettingsScreen()
+            }
         }
     }
 }
-
-data class BottomNavItem(val title: String, val route: String, val icon: ImageVector)
