+++ app/src/main/java/com/pavlov/nearWarSecrets/ui/setpassword/SetPasswordViewModel.kt
@@ -3,8 +3,11 @@ package com.pavlov.nearWarSecrets.ui.setpassword
 import androidx.lifecycle.ViewModel
 import com.pavlov.nearWarSecrets.util.APK
 import com.pavlov.nearWarSecrets.util.APKM
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
 
-class SetPasswordViewModel(
+@HiltViewModel
+class SetPasswordViewModel @Inject constructor(
     private val apkm: APKM
 ) : ViewModel() {
