+++ app/src/main/java/com/pavlov/nearWarSecrets/ui/theme/uiComponents/ButtonStyles.kt
@@ -1,5 +1,6 @@
 package com.pavlov.nearWarSecrets.ui.theme.uiComponents
 
+import com.pavlov.nearWarSecrets.R
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
