--- gradle/libs/libs.versions.toml
@@ -1,19 +1,0 @@
-[versions]
-coil = "2.7.0"
-compose = "1.7.5"
-hilt = "2.52"
-hiltNavigationCompose = "1.2.0"
-kotlin = "1.9.10"
-
-
-# Плагины (если необходимо)
-[plugins]
-android = { id = "com.android.application", version = "7.3.1" }
-
-# Библиотеки
-[libraries]
-coil-compose = { module = "io.coil-kt:coil-compose", version.ref = "coil" }
-compose-ui = { module = "androidx.compose.ui:ui", version.ref = "compose" }
-hilt-android = { group = "com.google.dagger", name = "hilt-android", version.ref = "hilt" }
-hilt-compiler = { group = "com.google.dagger", name = "hilt-compiler", version.ref = "hilt" }
-hilt-navigation-compose = { group = "androidx.hilt", name = "hilt-navigation-compose", version.ref = "hiltNavigationCompose" }
